move lock to utils::
