+++ storage/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod bitmap;
 pub mod bloom;
 pub mod tmpfile;
+pub mod lock;
